requirements.txt
setuptools
wheel
pandas==1.5.3
pyarrow
fastparquet
pyspark==3.3.0
delta-spark==2.1.0
psycopg2
sqlalchemy
python-dotenv
jupyter
flask
databricks-cli
streamlit
streamlit-echarts
streamlit-extras

azure-core==1.24.1
azure.identity
azure-storage-blob
azure-data-tables